fix(edge): address second-judgment round 1 on diagnostics accuracy
- State in the runbook that set_panic_endpoint moves panic output off stderr
  once the endpoint exists (SDK replaces the default hook), fails open before
  that, and that panic records carry no request ID so they correlate by
  timestamp/POP
- Document that x-divine-edge-request-id is trustworthy only once the outer
  VCL service is activated (req_id.rs and README)
- Move the persisted error-category mapping into blossom-core with a test
  asserting categories exist exactly for the variants whose status codes are
  5xx, so a remapped variant fails CI instead of silently logging null; the
  edge test binary cannot link locally, so the invariant must live in the
  crate CI executes
- Move the quality-variant predicate tests from the edge crate (compiled but
  never executed) into blossom-core and add transcript-path parse coverage
