+++ README.md
@@ -221,7 +221,7 @@ Viewer/list requests also accept NIP-98 HTTP auth (kind 27235) for the exact req
 
 ### Request correlation
 
-Admin and moderation endpoints accept an `X-Request-Id` header and echo it into related log lines (`[req=<id>]`) so retries and partial failures can be traced. When the outer Fastly VCL service is chained in front of Compute it pins the selected ID in `X-Divine-Edge-Request-Id`, which takes precedence; otherwise `X-Request-Id` is used. If neither is present, the leading segment of the Cloudflare `cf-ray` header is used; failing that, a short hex ID is generated. IDs are sanitized and capped at 64 characters, so common caller IDs (UUIDs, trace IDs) are preserved verbatim and longer values keep a 64-character prefix. Upstream services (e.g. the moderation service driving creator-delete) should forward a stable `X-Request-Id` across retries. `[PURGE]` logs carry the blob `sha256` as the surrogate key for cross-referencing cache purges.
+Admin and moderation endpoints accept an `X-Request-Id` header and echo it into related log lines (`[req=<id>]`) so retries and partial failures can be traced. When the outer Fastly VCL service is chained in front of Compute it pins the selected ID in `X-Divine-Edge-Request-Id`, which takes precedence and is overwritten on every chained request, so it is trustworthy only once that outer service is activated; until then a direct caller can set the header itself. Otherwise `X-Request-Id` is used. If neither is present, the leading segment of the Cloudflare `cf-ray` header is used; failing that, a short hex ID is generated. IDs are sanitized and capped at 64 characters, so common caller IDs (UUIDs, trace IDs) are preserved verbatim and longer values keep a 64-character prefix. Upstream services (e.g. the moderation service driving creator-delete) should forward a stable `X-Request-Id` across retries. `[PURGE]` logs carry the blob `sha256` as the surrogate key for cross-referencing cache purges.
 
 ## Deployment
