+++ blossom-core/src/types.rs
@@ -1276,6 +1276,84 @@ mod tests {
     }
 
     #[test]
+    fn quality_variant_path_valid() {
+        let hash = "a".repeat(64);
+        assert!(is_quality_variant_path(&format!("/{}/720p", hash)));
+        assert!(is_quality_variant_path(&format!("/{}/480p", hash)));
+        assert!(is_quality_variant_path(&format!("/{}/720p.mp4", hash)));
+        assert!(is_quality_variant_path(&format!("/{}/480p.mp4", hash)));
+
+        let (parsed_hash, filename, ct) =
+            parse_quality_variant_path(&format!("/{}/720p", hash)).unwrap();
+        assert_eq!(parsed_hash, hash);
+        assert_eq!(filename, "stream_720p.ts");
+        assert_eq!(ct, "video/mp2t");
+
+        let (parsed_hash, filename, ct) =
+            parse_quality_variant_path(&format!("/{}/720p.mp4", hash)).unwrap();
+        assert_eq!(parsed_hash, hash);
+        assert_eq!(filename, "stream_720p.mp4");
+        assert_eq!(ct, "video/mp4");
+    }
+
+    #[test]
+    fn quality_variant_path_no_underflow_on_short_input() {
+        // These must not panic (previously caused u32::MAX underflow)
+        assert!(!is_quality_variant_path("/720p"));
+        assert!(!is_quality_variant_path("/480p"));
+        assert!(!is_quality_variant_path("/720p.mp4"));
+        assert!(!is_quality_variant_path("/480p.mp4"));
+        assert!(!is_quality_variant_path("720p"));
+        assert!(!is_quality_variant_path("480p"));
+        assert!(!is_quality_variant_path(""));
+        assert!(parse_quality_variant_path("/480p").is_none());
+        assert!(parse_quality_variant_path("720p").is_none());
+        assert!(parse_quality_variant_path("/720p.mp4").is_none());
+        assert!(parse_quality_variant_path("480p.mp4").is_none());
+    }
+
+    #[test]
+    fn mp4_variant_maps_to_ts_counterpart() {
+        let hash = "a".repeat(64);
+
+        // 720p.mp4 derives correct .ts counterpart for backfill check
+        let (_, filename, ct) = parse_quality_variant_path(&format!("/{}/720p.mp4", hash)).unwrap();
+        assert_eq!(ct, "video/mp4");
+        assert_eq!(filename.replace(".mp4", ".ts"), "stream_720p.ts");
+
+        // 480p.mp4 likewise
+        let (_, filename, ct) = parse_quality_variant_path(&format!("/{}/480p.mp4", hash)).unwrap();
+        assert_eq!(ct, "video/mp4");
+        assert_eq!(filename.replace(".mp4", ".ts"), "stream_480p.ts");
+
+        // .ts variants have different content type — backfill path won't trigger
+        let (_, _, ct) = parse_quality_variant_path(&format!("/{}/720p", hash)).unwrap();
+        assert_eq!(ct, "video/mp2t");
+    }
+
+    #[test]
+    fn transcript_paths_parse_with_case_insensitive_suffix() {
+        let hash = "a".repeat(64);
+        for suffix in ["VTT", "vtt", "Vtt"] {
+            assert_eq!(
+                parse_transcript_path(&format!("/{hash}/{suffix}")),
+                Some(hash.clone()),
+                "suffix {suffix} must parse"
+            );
+        }
+        // Extra segments are not a transcript route.
+        assert_eq!(parse_transcript_path(&format!("/{hash}/vtt/x")), None);
+        assert_eq!(parse_transcript_path("/not-a-hash/vtt"), None);
+        // File form is extension-based, not segment-based.
+        assert_eq!(
+            parse_vtt_file_path(&format!("/{hash}.vtt")),
+            Some(hash.clone())
+        );
+        assert_eq!(parse_vtt_file_path("/not-a-hash.vtt"), None);
+        assert_eq!(parse_vtt_file_path(&format!("/{hash}.VTT")), None);
+    }
+
+    #[test]
     fn test_local_mode_stub_vtt_format() {
         let vtt = "WEBVTT\n\n00:00:00.000 --> 00:00:01.000\n[local mode stub transcript]\n";
         assert!(vtt.starts_with("WEBVTT"));
